+++ CHANGELOG.md
@@ -27,6 +27,7 @@ mPDF 8.0.x
 * Cache now creates a dedicated subdirectory `/mpdf`.
 * It is possible to disable automatic cache cleanup with `cacheCleanupInterval` config variable
 * PHP 8.0 is supported since 8.0.10 (#1263)
+* Fix: First header of named page is added twice (@antman3351, #1320)
 
 mPDF 8.0.0
 ===========================